addon_portal/requirements.txt
# Q2O Licensing Addon - Python Dependencies
# Compatible with Python 3.10-3.13

# Core Framework
fastapi>=0.110.0,<0.111.0
uvicorn[standard]>=0.27.0,<0.28.0

# Database
sqlalchemy>=2.0.25,<2.1.0
psycopg2-binary>=2.9.9,<3.0.0
alembic>=1.13.0,<1.14.0

# Data Validation
pydantic>=2.6.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0

# Payment Integration
stripe>=7.0.0,<8.0.0

# Authentication
pyjwt>=2.8.0,<3.0.0
Authlib>=1.3.0,<2.0.0

# Templates & Web
Jinja2>=3.1.3,<4.0.0
python-multipart>=0.0.6,<0.1.0

# Security
cryptography>=41.0.0,<42.0.0